+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## HEAD
 
+## 7.4.0
+
+- Change HerokuRun#call to use exceptions for empty output retries in preparation of related work
+- Thread safety for HerokuRun (https://github.com/heroku/hatchet/issues/157)
+- Timeout support for 'heroku run' (in case of networking, hanging or "server boot" tests etc), defaults to 60 s, global override `$HATCHET_DEFAULT_RUN_TIMEOUT`, per-test option `:timeout`
+- Delay for empty/timeout retries if run_multi is off, defaults to 1 s, global override `$HATCHET_RUN_RETRY_DELAY`
+- Record and print dyno id ("run.1234") in event of empty output or timeout retry
+- Terminate dyno via API on timeout
+- Allow returning of entire run object in App#run (to access process status) using :return_obj => true option
+
 ## 7.3.4
 
 - Memoize `Hatchet::App.default_buildpack` (https://github.com/heroku/hatchet/pull/183)